+++ sdk/keyvault/keyvault-keys/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 4.3.0-beta.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Key Bugs Fixed
+
+### Fixed
+
 ## 4.2.1 (2021-06-15)
 
 ### Bug Fixes